Integrated GPIO handling class to daemon/web server
Switched to using a derived class of HTTPServer that has _gpio (and
_zone for zone data class) members. Daemon class will create the GPIO
handler and pass it into the web server.Application will use these
instances during it's life cycle.

FIX: Noticed that atexit was not fiering properly when stopping the
daemon since atexit does not get called on SIGTERM. Added a SIGTERM
handler that gracefully terminate the daemon thus calling atexit.