Auto generate zone file on the first run
As not having the zone file when running the daemon for the first time
generated some confusion in the first alpha test, code was modified to
gracefully handle the creation of default zone data file in such cases.

Also including the default zone data file (zones.json-dist) to be usied
in packaging.