Finalizing Google Calendar integration
Google Calendar query will fetch and store the events from stat of the day, and
based on the current time it will update the "running" state of each event.

From the events stored in the schedule data, state of each sprinkler zone is
updated, and if the zone state is changed signal is set to the device via GPIO.

* Moved object creation to daemon class so the same instances will be used
through out the lifetime of the application.
* Moved the calendar id into data file (from config) and made it configurable
via web interface.
* Changed the API query delay to seconds instead of minutes. (Minimum/Default
to 10 seconds)