[bugfix] ISO time stamp filtering issue for GMT- time zones
Character '+' was used to identify the beginning of the time zone indicator.
This caused the GMT- time zones to be ignored in the filter.

Added code to also identify GMT- time zones using last '-' character when it is
found in the position larger than 10 (assuming that we have date component with
'-' separators with 'YYYY-MM-DD')