[bugfix] Zone doesn't turn off when schedule moved while it's running
*Cause*

Code was detecting event changes when it expired or deleted from the calendar.
Changing the event time is reflected in the local cache, but updating the
device (turning of the zone) was restricted because it would always zones that
was turned on manually. So the schedule can turn on a zone when schedule
adjustment falls with in the current time, but not wise versa.

*Solution*

* Introduced a new attribute to zone data structure to keep track of who (schedule or manually) changed the zone state (on/off).
* Changed the schedule thread to update the zone state regardless of if it's being turned on or off.
* Changed the statue update function on the zone storage so that when turning a zone by schedule is only possible if that zone was turned on by the scheduler.