Limit the duration a zone can run for.
Added a global limit (max_run - number of hours) to zone data file that the
application will allow a single zone to be turned on.

This limit affect zones both turned on via calendar and manually. Schedule
maintenance process checks the new events for duration and only record and
enable events where the duration is less or equal to the max_run limit.