[bugfix] Time adjustment in calender doesn't turn zone on
*Issue*

When a event scheduled for a future time is moved in to slot that falls within
the current time, it doesn't turn the zone on.

*Cause*

Fix that was done to make zone turn off manually after it was started via
calender also caused this scenario to fall it to the same category and ignore
the turn on.

*Solution*

Instead of determining if the zone was manually turned off using the date/time
and state_owner, introduced a new flag to indicate that zone turned on by
calendar was manually turned off (manual_off).